requirements.txt
Django==1.11.1
django-crispy-forms==1.7.2
django-model-utils==3.1.1
django-notifications-hq==1.3
jsonfield==2.0.2
Pillow==5.1.0
pkg-resources==0.0.0
pytz==2018.4
six==1.11.0
transliterate==1.10.1
